requirements.txt
Flask
Flask-SQLAlchemy
pygments
nose
coverage
gunicorn
